requirements.txt
streamlit
streamlit-cropper
mediapipe
opencv-python-headless
Pillow
pdf2image
numpy